# Speed up Integer#digit