```
Warming up --------------------------------------
    digit_1 (legacy)    60.395k i/100ms
    digit_2 (v0.1.0)    70.661k i/100ms
             digit_3    73.059k i/100ms
             digit_4    77.495k i/100ms
Calculating -------------------------------------
    digit_1 (legacy)      1.001M (± 5.5%) i/s -      5.013M in   5.023945s
    digit_2 (v0.1.0)      1.185M (± 6.4%) i/s -      5.936M in   5.031353s
             digit_3      1.270M (±10.4%) i/s -      6.283M in   5.050293s
             digit_4      1.468M (± 8.3%) i/s -      7.285M in   5.007498s

Comparison:
             digit_4:  1467584.4 i/s
             digit_3:  1269908.5 i/s - same-ish: difference falls within error
    digit_2 (v0.1.0):  1184810.6 i/s - 1.24x slower
    digit_1 (legacy):  1000996.2 i/s - 1.47x slower

```